feat(process): offload scanning to a worker thread
This reimplements the original arRPC PR 133 from pendo324. But with changes from 109 (#3), 111 (#4) and 123 (#5) also included.
